Update core - Minor changes.
