+++ core/src/processor/cuda/CUDAProcessor.cpp
@@ -16,6 +16,8 @@
 #include "AC/Core/Util.hpp"
 #include "AC/Util/ThreadLocal.hpp"
 
+#include "AC/Core/Internal/Processor/CUDA/Common.hpp"
+
 #include "ACCoreExport.hpp" // Generated by CMake
 
 #define ContextList (ac::core::cuda::getContextList())
@@ -265,36 +267,6 @@ namespace ac::core::cuda
         return contextList;
     }
 
-    struct DeviceImage
-    {
-        void* ptr = nullptr;
-        int w = 0;
-        int h = 0;
-        int c = 0;
-        int pitch = 0;
-        int elementSize = 0;
-
-        void create(const int w, const int h, const int c, const int elementSize) noexcept
-        {
-            this->w = w;
-            this->h = h;
-            this->c = c;
-            this->elementSize = elementSize;
-            pitch = align(w * c * elementSize, 128);
-        }
-
-        cudaError_t fromHost(const Image& hostImage, const cudaStream_t stream) const noexcept
-        {
-            auto lineSize = hostImage.width() * hostImage.pixelSize();
-            return cudaMemcpy2DAsync(ptr, pitch, hostImage.ptr(), hostImage.stride(), lineSize, hostImage.height(), cudaMemcpyHostToDevice, stream);
-        }
-        cudaError_t toHost(Image& hostImage, const cudaStream_t stream) const noexcept
-        {
-            auto lineSize = w * c * elementSize;
-            return cudaMemcpy2DAsync(hostImage.ptr(), hostImage.stride(), ptr, pitch, lineSize, h, cudaMemcpyDeviceToHost, stream);
-        }
-    };
-
     class DeviceImageAllocator
     {
     public:
@@ -497,32 +469,32 @@ namespace ac::core::cuda
             for (int i = 0; i < model.kernels(); i++)
             {
                 int layer = model.kernelLayer(i);
-                float* dkptr = nullptr;
-                auto hkptr = model.kernel(layer);
+                float* dptr = nullptr;
+                auto hptr = model.kernel(layer);
                 auto size = model.kernelSize(layer);
-                err = cudaMalloc(&dkptr, size); if (err != cudaSuccess) return;
-                kernels.emplace_back(dkptr);
-                err = cudaMemcpy(dkptr, hkptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
+                err = cudaMalloc(&dptr, size); if (err != cudaSuccess) return;
+                kernels.emplace_back(dptr);
+                err = cudaMemcpy(dptr, hptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
             }
             for (int i = 0; i < model.biases(); i++)
             {
                 int layer = model.biasLayer(i);
-                float* dkptr = nullptr;
-                auto hkptr = model.bias(layer);
+                float* dptr = nullptr;
+                auto hptr = model.bias(layer);
                 auto size = model.biasSize(layer);
-                err = cudaMalloc(&dkptr, size); if (err != cudaSuccess) return;
-                biases.emplace_back(dkptr);
-                err = cudaMemcpy(dkptr, hkptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
+                err = cudaMalloc(&dptr, size); if (err != cudaSuccess) return;
+                biases.emplace_back(dptr);
+                err = cudaMemcpy(dptr, hptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
             }
             for (int i = 0; i < model.alphas(); i++)
             {
                 int layer = model.alphaLayer(i);
-                float* dkptr = nullptr;
-                auto hkptr = model.alpha(layer);
+                float* dptr = nullptr;
+                auto hptr = model.alpha(layer);
                 auto size = model.alphaSize(layer);
-                err = cudaMalloc(&dkptr, size); if (err != cudaSuccess) return;
-                alphas.emplace_back(dkptr);
-                err = cudaMemcpy(dkptr, hkptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
+                err = cudaMalloc(&dptr, size); if (err != cudaSuccess) return;
+                alphas.emplace_back(dptr);
+                err = cudaMemcpy(dptr, hptr, size, cudaMemcpyHostToDevice); if (err != cudaSuccess) return;
             }
         }
         ~CUDAProcessorSeqCNN() noexcept override
